ParseSrdSorts/requirements.txt
azure-functions
azure-storage-blob
pdfplumber
PyPDF2